+++ operators/__init__.py
@@ -1,0 +1,1 @@
+"""Blender operators for LEGO Builder for Blender."""